chore(radio-traffic): remove the RadioScanner shell, document the mp3 API (closes #020-7f21)
RadioTraffic replaced RadioScanner's UI weeks ago; this deletes what was
left of the old shell (RadioScanner.tsx, RadioScannerContext.ts,
NowPlayingList.tsx + tests) after moving radioTuneStation into
radio-core/ so PlaylistProvider's import survives the deletion, and
swaps the Desktop.tsx app registration accordingly. Also documents the
mp3 public metadata columns and mp3_tags/mp3_items_tags in
data-model.md, the rederive-mp3-metadata flow in party-identification.md
(websocket-protocol.md and http-api.md were already complete), and
drops the dead TAG_INDEX_SQL from seed.mjs (it indexed a mp3_items.tags
column that no longer exists now that tags is an m2m alias).
