+++ packages/frontend/CLAUDE.md
@@ -9,7 +9,7 @@ Guidance for AI coding assistants working in this Vite + React + TypeScript app.
 The Mac OS 8-style desktop shell for 911realtime.org. It is built on the [`classicy`](https://www.npmjs.com/package/classicy) npm package — an external, generically-reusable retro desktop/window-manager component library. Classicy ships its own bundled system apps (Finder, PDF Viewer, Picture Viewer, Movie Player, SimpleText, Control Panels) — **this repo never implements those**; it only supplies configuration and content for them. What this package *does* own:
 
 - `src/app.tsx` — the root: wraps everything in `ClassicyAppManagerProvider` (default file system, seeded boot state) and `MediaStreamProvider`, then renders each Application inside `ClassicyDesktop`.
-- `src/Applications/*` — one folder per product app (Browser, News, TV, RadioScanner, PagerDecoder, Newsgroups, TimeMachine, Feedback, Account), each a `ClassicyApp` + `ClassicyWindow(s)`.
+- `src/Applications/*` — one folder per product app (Browser, News, TV, RadioTraffic, RadioTuner, PagerDecoder, Newsgroups, TimeMachine, Feedback, Account), each a `ClassicyApp` + `ClassicyWindow(s)`. `src/Applications/radio-core/` is not an app — it's the shared library (station grouping/audio pipeline/players/settings) that RadioTraffic, RadioTuner, PlaylistEditor, MarketWatch, Mobile and `Providers/Playlist` all build on; the dependency arrow points one way (apps depend on radio-core, never the reverse — enforced by `radio-core/noAppImports.test.ts`). It's what's left after the old RadioScanner app was split into RadioTraffic (comm/ATC traffic) and RadioTuner (continuous broadcast stations) with their shared code lifted out first.
 - `src/Providers/MediaStream/` — the single WebSocket client to the `packages/backend` streamer.
 - `src/Providers/Playlist/` — the teacher-playlist engine (spec: `plans/2026-07-16-teacher-playlists-design.md`): loads `?playlist=<id>` from Directus and enforces availability windows / app disabling / focus / scheduled events. Its runtime state is deliberately non-persisted (no ClassicyStore, no localStorage, no ClassicyFileSystem). `RoomControlBridge.tsx` sits alongside it and applies a teacher's **live** commands (jump/focus/message/lock), which the streamer pushes over the WebSocket to everyone following the same `?playlist=` id — see "Room control" in the backend's `docs/websocket-protocol.md`. The teacher end is a `Control` menu on each playlist document window in the Playlists app (`Applications/PlaylistEditor/playlistMenus.ts`'s `documentControlMenu`, wired up in `PlaylistDocumentWindow.tsx`), which POSTs to the streamer's `/room` via `roomApi.ts`; the streamer authorises those per playlist, so only the playlist's creator can drive it.
 - `src/Providers/Auth/` — Directus session auth (AuthProvider) + the playlistApi editor seam; auth state is never persisted client-side (httpOnly cookie only).
@@ -20,7 +20,7 @@ The Mac OS 8-style desktop shell for 911realtime.org. It is built on the [`class
 
 ## Mental model — read before changing anything
 
-- **One virtual clock, one writer seam.** The canonical time lives in Classicy's own state at `state.System.Manager.DateAndTime` (`{ dateTime, timeZoneOffset }`), seeded in `app.tsx` to `2001-09-11T12:40:00.000Z` / `-4`. **Every clock write goes through `setDateTimeFromUtc` in `TimeMachine/setVirtualClock.ts` — that's the invariant, not a fixed list of callers.** Three sanctioned writers use it today: `TimeMachine.tsx` (desktop) and the `src/Mobile/screens/` Time Travel screens (mobile) are the **user-driven** seams (the two are never mounted together); the Playlist engine drives **playlist jumps** — both the pre-authored kind (`src/Providers/Playlist/PlaylistProvider.tsx`, scheduled `jump` entries) and the **live teacher kind** (`src/Providers/Playlist/RoomControlBridge.tsx`, a `room_command` with action `jump` relayed by the streamer to everyone in the room); and `MediaStreamProvider` drives the clock **while the streamer's forced clock mode is active** (operator-forced broadcast time — see `docs/websocket-protocol.md`'s "Forced clock mode" section on the backend), snapping to the server's `clock`/`heartbeat_ack.master_time` frames and locking `DateAndTime` so other writers can't fight it. Every other app — `TV`, `RadioScanner`, `News`, `PagerDecoder` — reads the clock reactively via `useClassicyDateTime({ tick: true })` (when it needs play/pause/`localDate`) or `useAppManager((s) => s.System.Manager.DateAndTime.dateTime)` (when it just needs the value). Don't add a new place that calls `setDateTime`/`shiftTime` directly — route through the seam instead.
+- **One virtual clock, one writer seam.** The canonical time lives in Classicy's own state at `state.System.Manager.DateAndTime` (`{ dateTime, timeZoneOffset }`), seeded in `app.tsx` to `2001-09-11T12:40:00.000Z` / `-4`. **Every clock write goes through `setDateTimeFromUtc` in `TimeMachine/setVirtualClock.ts` — that's the invariant, not a fixed list of callers.** Three sanctioned writers use it today: `TimeMachine.tsx` (desktop) and the `src/Mobile/screens/` Time Travel screens (mobile) are the **user-driven** seams (the two are never mounted together); the Playlist engine drives **playlist jumps** — both the pre-authored kind (`src/Providers/Playlist/PlaylistProvider.tsx`, scheduled `jump` entries) and the **live teacher kind** (`src/Providers/Playlist/RoomControlBridge.tsx`, a `room_command` with action `jump` relayed by the streamer to everyone in the room); and `MediaStreamProvider` drives the clock **while the streamer's forced clock mode is active** (operator-forced broadcast time — see `docs/websocket-protocol.md`'s "Forced clock mode" section on the backend), snapping to the server's `clock`/`heartbeat_ack.master_time` frames and locking `DateAndTime` so other writers can't fight it. Every other app — `TV`, `RadioTraffic`, `RadioTuner`, `News`, `PagerDecoder` — reads the clock reactively via `useClassicyDateTime({ tick: true })` (when it needs play/pause/`localDate`) or `useAppManager((s) => s.System.Manager.DateAndTime.dateTime)` (when it just needs the value). Don't add a new place that calls `setDateTime`/`shiftTime` directly — route through the seam instead.
 - **One WebSocket, ref-counted subscriptions.** `MediaStreamProvider` opens exactly one connection (`VITE_MEDIA_STREAM_URL`) in `app.tsx`, above `ClassicyDesktop`. Every app shares it through `MediaStreamContext` — either directly (`useContext(MediaStreamContext)`) or via the filtering hook `useMediaStream(filter)`. Opt-in side channels (pager/mp3/news/usenet) are tracked as a `Set<appId>` per channel so multiple apps can subscribe/unsubscribe independently without duplicating `{type:"subscribe"}` wire traffic — see the `subscribePager`/`unsubscribePager`-style pairs in `MediaStreamProvider.tsx`.
 - **Incoming data is buffered, not applied immediately.** Frames land in per-channel `Map` buffers and a per-second effect promotes items whose `start_date` has arrived (`revealBuffer.ts`) and prunes expired ones (`retention.ts`) — this is what makes playback advance in lockstep with the virtual clock instead of dumping the whole dataset at once. A clock jump past `SEEK_THRESHOLD_MS` (90s) is treated as a manual seek: buffers clear and a `{type:"seek"}` message goes out.
 - **`virtualUtcMs` strips the display timezone back off.** `useClassicyDateTime`'s `localDate` is a *display* value (shifted for the menu-bar clock); the streamer, item `start_date`s, and seek/heartbeat messages are all true UTC. `Providers/MediaStream/virtualClock.ts`'s `virtualUtcMs(localDate, tzOffsetHours)` recovers the real UTC instant. Comparing `localDate` directly against wire timestamps previously trapped short-lived items (radio clips, instant news) permanently in the reveal buffer for any non-zero offset — see `virtualClock.test.ts`.
