+++ packages/frontend/src/Providers/Playlist/PlaylistProvider.tsx
@@ -25,8 +25,8 @@ import {
 import { browserNavigate } from "../../Applications/Browser/BrowserContext";
 import { flightTrackerFocusFlight } from "../../Applications/FlightTracker/flightTrackerCommands";
 import { newsFocusItem } from "../../Applications/News/NewsContext";
-import { radioTuneStation } from "../../Applications/RadioScanner/RadioScannerContext";
 import { BROADCAST_STATIONS } from "../../Applications/radio-core/stationGrouping";
+import { radioTuneStation } from "../../Applications/radio-core/radioTuneStation";
 import { radioTunerTuneStation } from "../../Applications/RadioTuner/RadioTunerContext";
 import { setDateTimeFromUtc } from "../../Applications/TimeMachine/setVirtualClock";
 import { tvTuneChannel } from "../../Applications/TV/TVContext";
